Move DataServices/Models to DataServices.Models; remove superfluous ToJson methods from Models
