+++ src/DataServices/DataServices.Models/ReturnCode.cs
@@ -9,18 +9,35 @@
  */
 
 using System;
+using System.ComponentModel.DataAnnotations;
 using System.Runtime.Serialization;
 using System.Text;
-using Newtonsoft.Json;
+using System.Xml;
 
 namespace DataServices.Models
 { 
     /// <summary>
-    /// The unique identifier for the SDRA assessment document i.e. the sdoc_id
+    /// 
     /// </summary>
     [DataContract]
-    public partial class SdocId : IEquatable<SdocId>
+    public partial class ReturnCode : IEquatable<ReturnCode>
     { 
+        /// <summary>
+        /// The &#x27;error&#x27; code number returned by SDRA webservice
+        /// </summary>
+        /// <value>The &#x27;error&#x27; code number returned by SDRA webservice</value>
+        [Required]
+        [DataMember(Name="code")]
+        public int? Code { get; set; }
+
+        /// <summary>
+        /// The description of the &#x27;error&#x27; or state
+        /// </summary>
+        /// <value>The description of the &#x27;error&#x27; or state</value>
+        [Required]
+        [DataMember(Name="message")]
+        public string Message { get; set; }
+
         /// <summary>
         /// Returns the string presentation of the object
         /// </summary>
@@ -28,18 +45,11 @@ namespace DataServices.Models
         public override string ToString()
         {
             var sb = new StringBuilder();
-            sb.Append("class SdocId {\n");
+            sb.Append("class ReturnCode {\n");
+            sb.Append("  Code: ").Append(Code).Append("\n");
+            sb.Append("  Message: ").Append(Message).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
         }
 
         /// <summary>
@@ -51,20 +61,30 @@ namespace DataServices.Models
         {
             if (ReferenceEquals(null, obj)) return false;
             if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((SdocId)obj);
+            return obj.GetType() == GetType() && Equals((ReturnCode)obj);
         }
 
         /// <summary>
-        /// Returns true if SdocId instances are equal
+        /// Returns true if ReturnCode instances are equal
         /// </summary>
-        /// <param name="other">Instance of SdocId to be compared</param>
+        /// <param name="other">Instance of ReturnCode to be compared</param>
         /// <returns>Boolean</returns>
-        public bool Equals(SdocId other)
+        public bool Equals(ReturnCode other)
         {
             if (ReferenceEquals(null, other)) return false;
             if (ReferenceEquals(this, other)) return true;
 
-            return false;
+            return 
+                (
+                    Code == other.Code ||
+                    Code != null &&
+                    Code.Equals(other.Code)
+                ) && 
+                (
+                    Message == other.Message ||
+                    Message != null &&
+                    Message.Equals(other.Message)
+                );
         }
 
         /// <summary>
@@ -77,6 +97,10 @@ namespace DataServices.Models
             {
                 var hashCode = 41;
                 // Suitable nullity checks etc, of course :)
+                    if (Code != null)
+                    hashCode = hashCode * 59 + Code.GetHashCode();
+                    if (Message != null)
+                    hashCode = hashCode * 59 + Message.GetHashCode();
                 return hashCode;
             }
         }
@@ -84,12 +108,12 @@ namespace DataServices.Models
         #region Operators
         #pragma warning disable 1591
 
-        public static bool operator ==(SdocId left, SdocId right)
+        public static bool operator ==(ReturnCode left, ReturnCode right)
         {
             return Equals(left, right);
         }
 
-        public static bool operator !=(SdocId left, SdocId right)
+        public static bool operator !=(ReturnCode left, ReturnCode right)
         {
             return !Equals(left, right);
         }
