+++ src/DataServices/DataServices/Controllers/AssessmentApi.cs
@@ -8,17 +8,16 @@
  * Generated by: https://github.com/swagger-api/swagger-codegen.git
  */
 
-using DataServices.Attributes;
-using DataServices.Models;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-using Swashbuckle.AspNetCore.Annotations;
 using System;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using DataServices.Adapters;
+using DataServices.Attributes;
 using DataServices.Connected_Services.SDRAAssessmentWebService;
+using DataServices.Models;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
+using Swashbuckle.AspNetCore.Annotations;
 using Document = DataServices.Connected_Services.SDRAAssessmentWebService.Document;
 
 namespace DataServices.Controllers
