--- src/DataServices/DataServices/Models/QueuedDocumentObject.cs
@@ -1,160 +1,0 @@
-/*
- * SDRA API
- *
- * This API is for SDRA  It provides Source Document Assessment Data and Data Access 
- *
- * OpenAPI spec version: 1.0.0
- * 
- * Generated by: https://github.com/swagger-api/swagger-codegen.git
- */
-
-using System;
-using System.ComponentModel.DataAnnotations;
-using System.Runtime.Serialization;
-using System.Text;
-using Newtonsoft.Json;
-
-namespace DataServices.Models
-{ 
-    /// <summary>
-    /// 
-    /// </summary>
-    [DataContract]
-    public partial class QueuedDocumentObject : IEquatable<QueuedDocumentObject>
-    { 
-        /// <summary>
-        /// Gets or Sets SodcId
-        /// </summary>
-        [DataMember(Name="sodcId")]
-        public SdocId SodcId { get; set; }
-
-        /// <summary>
-        /// The &#x27;error&#x27; code number returned by SDRA webservice
-        /// </summary>
-        /// <value>The &#x27;error&#x27; code number returned by SDRA webservice</value>
-        [DataMember(Name="code")]
-        public int? Code { get; set; }
-
-        /// <summary>
-        /// Gets or Sets Message
-        /// </summary>
-        [Required]
-        [DataMember(Name="message")]
-        public string Message { get; set; }
-
-        /// <summary>
-        /// Gets or Sets StatusTime
-        /// </summary>
-        [Required]
-        [DataMember(Name="statusTime")]
-        public DateTime? StatusTime { get; set; }
-
-        /// <summary>
-        /// Returns the string presentation of the object
-        /// </summary>
-        /// <returns>String presentation of the object</returns>
-        public override string ToString()
-        {
-            var sb = new StringBuilder();
-            sb.Append("class QueuedDocumentObject {\n");
-            sb.Append("  SodcId: ").Append(SodcId).Append("\n");
-            sb.Append("  Code: ").Append(Code).Append("\n");
-            sb.Append("  Message: ").Append(Message).Append("\n");
-            sb.Append("  StatusTime: ").Append(StatusTime).Append("\n");
-            sb.Append("}\n");
-            return sb.ToString();
-        }
-
-        /// <summary>
-        /// Returns the JSON string presentation of the object
-        /// </summary>
-        /// <returns>JSON string presentation of the object</returns>
-        public string ToJson()
-        {
-            return JsonConvert.SerializeObject(this, Formatting.Indented);
-        }
-
-        /// <summary>
-        /// Returns true if objects are equal
-        /// </summary>
-        /// <param name="obj">Object to be compared</param>
-        /// <returns>Boolean</returns>
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj)) return false;
-            if (ReferenceEquals(this, obj)) return true;
-            return obj.GetType() == GetType() && Equals((QueuedDocumentObject)obj);
-        }
-
-        /// <summary>
-        /// Returns true if QueuedDocumentObject instances are equal
-        /// </summary>
-        /// <param name="other">Instance of QueuedDocumentObject to be compared</param>
-        /// <returns>Boolean</returns>
-        public bool Equals(QueuedDocumentObject other)
-        {
-            if (ReferenceEquals(null, other)) return false;
-            if (ReferenceEquals(this, other)) return true;
-
-            return 
-                (
-                    SodcId == other.SodcId ||
-                    SodcId != null &&
-                    SodcId.Equals(other.SodcId)
-                ) && 
-                (
-                    Code == other.Code ||
-                    Code != null &&
-                    Code.Equals(other.Code)
-                ) && 
-                (
-                    Message == other.Message ||
-                    Message != null &&
-                    Message.Equals(other.Message)
-                ) && 
-                (
-                    StatusTime == other.StatusTime ||
-                    StatusTime != null &&
-                    StatusTime.Equals(other.StatusTime)
-                );
-        }
-
-        /// <summary>
-        /// Gets the hash code
-        /// </summary>
-        /// <returns>Hash code</returns>
-        public override int GetHashCode()
-        {
-            unchecked // Overflow is fine, just wrap
-            {
-                var hashCode = 41;
-                // Suitable nullity checks etc, of course :)
-                    if (SodcId != null)
-                    hashCode = hashCode * 59 + SodcId.GetHashCode();
-                    if (Code != null)
-                    hashCode = hashCode * 59 + Code.GetHashCode();
-                    if (Message != null)
-                    hashCode = hashCode * 59 + Message.GetHashCode();
-                    if (StatusTime != null)
-                    hashCode = hashCode * 59 + StatusTime.GetHashCode();
-                return hashCode;
-            }
-        }
-
-        #region Operators
-        #pragma warning disable 1591
-
-        public static bool operator ==(QueuedDocumentObject left, QueuedDocumentObject right)
-        {
-            return Equals(left, right);
-        }
-
-        public static bool operator !=(QueuedDocumentObject left, QueuedDocumentObject right)
-        {
-            return !Equals(left, right);
-        }
-
-        #pragma warning restore 1591
-        #endregion Operators
-    }
-}
